server: Detach socket io process
